trim fix

+++ SNDG/Comparative/Mapping.py
@@ -50,10 +50,13 @@ class Mapping:
         e("fastqc trimmed_1.fastq", work_dir)
         e("fastqc trimmed_2.fastq", work_dir)
 
-        e("cat trimmed_1_singletons.fastq >> trimmed_s.fastq", work_dir)
-        e("cat trimmed_2_singletons.fastq >> trimmed_s.fastq", work_dir)
-        os.remove(work_dir + "trimmed_1_singletons.fastq")
-        os.remove(work_dir + "trimmed_2_singletons.fastq")
+        if not os.path.exists(work_dir  + "trimmed_1_singletons.fastq"):
+            e("cat trimmed_1_singletons.fastq >> trimmed_s.fastq", work_dir)
+            os.remove(work_dir + "trimmed_1_singletons.fastq")
+
+        if not os.path.exists(work_dir  + "trimmed_2_singletons.fastq"):
+            e("cat trimmed_2_singletons.fastq >> trimmed_s.fastq", work_dir)
+            os.remove(work_dir + "trimmed_2_singletons.fastq")
 
     @staticmethod
     def alignment(work_dir, record, trimmed_1="trimmed_1.fastq",
